#185 feat: TripFilterProvider  여행지 필터 객체 생성
