+++ backend/src/main/java/moheng/recommendtrip/application/RecommendTripService.java
@@ -1,58 +1,45 @@
 package moheng.recommendtrip.application;
 
-import moheng.keyword.domain.TripKeyword;
 import moheng.keyword.domain.TripKeywordRepository;
-import moheng.liveinformation.domain.LiveInformation;
-import moheng.liveinformation.domain.MemberLiveInformationRepository;
-import moheng.liveinformation.domain.TripLiveInformationRepository;
 import moheng.member.domain.Member;
 import moheng.member.domain.repository.MemberRepository;
 import moheng.member.exception.NoExistMemberException;
 import moheng.recommendtrip.domain.RecommendTrip;
+import moheng.recommendtrip.domain.TripFilterProvider;
 import moheng.recommendtrip.domain.RecommendTripRepository;
 import moheng.recommendtrip.dto.RecommendTripCreateRequest;
 import moheng.recommendtrip.exception.LackOfRecommendTripException;
 import moheng.recommendtrip.exception.NoExistMemberTripException;
 import moheng.trip.domain.*;
 import moheng.trip.dto.FindTripsResponse;
-import moheng.trip.dto.RecommendTripsByVisitedLogsRequest;
-import moheng.trip.dto.RecommendTripsByVisitedLogsResponse;
-import moheng.trip.exception.NoExistRecommendTripException;
 import moheng.trip.exception.NoExistTripException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.*;
-import java.util.stream.Collectors;
 
 @Transactional(readOnly = true)
 @Service
 public class RecommendTripService {
     private static final int RECOMMEND_TRIPS_COUNT = 10;
     private static final int MIN_RECOMMEND_TRIPS_COUNT = 5;
-    private final ExternalRecommendModelClient externalRecommendModelClient;
+    private final TripFilterProvider tripFilterProvider;
     private final RecommendTripRepository recommendTripRepository;
     private final MemberRepository memberRepository;
     private final TripRepository tripRepository;
-    private final TripLiveInformationRepository tripLiveInformationRepository;
-    private final MemberLiveInformationRepository memberLiveInformationRepository;
     private final MemberTripRepository memberTripRepository;
     private final TripKeywordRepository tripKeywordRepository;
 
-    public RecommendTripService(final ExternalRecommendModelClient externalRecommendModelClient,
+    public RecommendTripService(final TripFilterProvider tripFilterProvider,
                                 final RecommendTripRepository recommendTripRepository,
                                 final MemberRepository memberRepository,
                                 final TripRepository tripRepository,
-                                final TripLiveInformationRepository tripLiveInformationRepository,
-                                final MemberLiveInformationRepository memberLiveInformationRepository,
                                 final MemberTripRepository memberTripRepository,
                                 final TripKeywordRepository tripKeywordRepository) {
-        this.externalRecommendModelClient = externalRecommendModelClient;
+        this.tripFilterProvider = tripFilterProvider;
         this.recommendTripRepository = recommendTripRepository;
         this.memberRepository = memberRepository;
         this.tripRepository = tripRepository;
-        this.tripLiveInformationRepository = tripLiveInformationRepository;
-        this.memberLiveInformationRepository = memberLiveInformationRepository;
         this.memberTripRepository = memberTripRepository;
         this.tripKeywordRepository = tripKeywordRepository;
     }
@@ -63,29 +50,8 @@ public class RecommendTripService {
         final List<MemberTrip> memberTrips = memberTripRepository.findByMember(member);
         final List<RecommendTrip> recommendTrips = recommendTripRepository.findTop10ByMember(member);
         final Map<Long, Long> preferredLocations = findMemberPreferredLocations(memberTrips, recommendTrips);
-        final List<Trip> filteredTrips = findFilteredTrips(preferredLocations, memberId);
+        final List<Trip> filteredTrips = tripFilterProvider.findFilteredTrips(preferredLocations, memberId);
         return new FindTripsResponse(tripKeywordRepository.findByTrips(filteredTrips));
-    }
-
-    private List<Trip> findFilteredTrips(final Map<Long, Long> preferredLocations, final long memberId) {
-        final List<Trip> filteredTrips = new ArrayList<>();
-        long page = 1L;
-        while (filteredTrips.size() < RECOMMEND_TRIPS_COUNT) {
-            final List<Trip> filteredTripsByLiveinformation = findRecommendTripsByModelClient(preferredLocations, memberId, page);
-            filteredTrips.addAll(filteredTripsByLiveinformation);
-            page++;
-        }
-        if (filteredTrips.size() > RECOMMEND_TRIPS_COUNT) {
-            return filteredTrips.subList(0, RECOMMEND_TRIPS_COUNT);
-        }
-        return filteredTrips;
-    }
-
-    private List<Trip> findRecommendTripsByModelClient(final Map<Long, Long> preferredLocations, final long memberId, final long page) {
-        final RecommendTripsByVisitedLogsResponse response = externalRecommendModelClient.recommendTripsByVisitedLogs(
-                new RecommendTripsByVisitedLogsRequest(preferredLocations, page)
-        );
-        return filterTripsByLiveinformation(response, memberId);
     }
 
     private Map<Long, Long> findMemberPreferredLocations(final List<MemberTrip> memberTrips, final List<RecommendTrip> recommendTrips) {
@@ -107,19 +73,6 @@ public class RecommendTripService {
             throw new LackOfRecommendTripException("추천을 받기위한 선호 여행지 데이터 수가 부족합니다.");
         }
     }
-
-    private List<Trip> filterTripsByLiveinformation(final RecommendTripsByVisitedLogsResponse recommendTripsByVisitedLogsResponse, final Long memberId) {
-        final List<Trip> trips = tripRepository.findTripsByContentIds(recommendTripsByVisitedLogsResponse.getContentIds());
-        final List<LiveInformation> memberLiveInformations = memberLiveInformationRepository.findLiveInformationsByMemberId(memberId);
-        return filterTripsByMemberInformation(trips, memberLiveInformations);
-    }
-
-    private List<Trip> filterTripsByMemberInformation(final List<Trip> trips, final List<LiveInformation> memberLiveInformations) {
-        return trips.stream()
-                .filter(trip -> tripLiveInformationRepository.existsByTripAndLiveInformationIn(trip, memberLiveInformations))
-                .collect(Collectors.toList());
-    }
-
 
     @Transactional
     public void saveByRank(final Trip trip, final Member member, long rank) {
